Move components around - flat
