+++ client/components/Collapse.css
@@ -1,0 +1,7 @@
+@import '../style/colors.css';
+@import '../style/variables.css';
+
+.collapse {
+  list-style-type: none;
+  padding: 0;
+}
