--- client/components/Collapse/index.js
@@ -1,2 +1,0 @@
-import Collapse from "./Collapse";
-export default Collapse;
