+++ client/components/Content.css
@@ -1,5 +1,5 @@
-@import '../../style/colors.css';
-@import '../../style/variables.css';
+@import '../style/colors.css';
+@import '../style/variables.css';
 
 :root {
   --content-padding-left: 20px;
