+++ client/components/Header.css
@@ -1,5 +1,5 @@
-@import '../../style/colors.css';
-@import '../../style/variables.css';
+@import '../style/colors.css';
+@import '../style/variables.css';
 
 :root {
   --header-height: 40px;
