--- client/components/Header/index.js
@@ -1,2 +1,0 @@
-import Header from "./Header";
-export default Header;
