--- client/components/PrettyPrint/index.js
@@ -1,2 +1,0 @@
-import PrettyPrint from './PrettyPrint';
-export default PrettyPrint;
