--- client/components/Reporter/index.js
@@ -1,2 +1,0 @@
-import Reporter from "./Reporter";
-export default Reporter;
